adding some unit tests
