+++ package.json
@@ -8,7 +8,8 @@
     "build": "tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
-    "back": "json-server --watch src/db/images.json --port 3000"
+    "back": "json-server --watch src/db/images.json --port 3000",
+    "test": "vitest --ui"
   },
   "dependencies": {
     "@auth0/auth0-react": "^2.2.1",
@@ -20,13 +21,15 @@
     "@tanstack/react-query-devtools": "^4.36.1",
     "axios": "^1.5.1",
     "bootstrap": "^5.3.2",
+    "jsdom": "^22.1.0",
     "json-server": "^0.17.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-image-crop": "^10.1.8",
     "react-image-file-resizer": "^0.4.8",
     "react-router-dom": "^6.16.0",
-    "reactstrap": "^9.2.0"
+    "reactstrap": "^9.2.0",
+    "vitest": "^0.34.6"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.1.3",
@@ -37,6 +40,7 @@
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
     "@vitejs/plugin-react": "^4.0.3",
+    "@vitest/ui": "^0.34.6",
     "eslint": "^8.45.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
